+++ packages/office-ui-fabric-react/src/components/DocumentCard/DocumentCardStatus.ts
@@ -3,9 +3,8 @@ import { DocumentCardStatusBase } from './DocumentCardStatus.base';
 import { getStyles } from './DocumentCardStatus.styles';
 import { IDocumentCardStatusProps, IDocumentCardStatusStyleProps, IDocumentCardStatusStyles } from './DocumentCardStatus.types';
 
-export const DocumentCardStatus = styled<IDocumentCardStatusProps, IDocumentCardStatusStyleProps, IDocumentCardStatusStyles>(
-  DocumentCardStatusBase,
-  getStyles,
-  undefined,
-  { scope: 'DocumentCardStatus' }
-);
+export const DocumentCardStatus: React.StatelessComponent<IDocumentCardStatusProps> = styled<
+  IDocumentCardStatusProps,
+  IDocumentCardStatusStyleProps,
+  IDocumentCardStatusStyles
+>(DocumentCardStatusBase, getStyles, undefined, { scope: 'DocumentCardStatus' });