+++ packages/office-ui-fabric-react/src/components/Shimmer/Shimmer.types.ts
@@ -55,36 +55,45 @@ export interface IShimmerProps extends React.AllHTMLAttributes<HTMLElement> {
    * Theme provided by High-Order Component.
    */
   theme?: ITheme;
+
+  /**
+   * Defines an object with possible colors to pass for Shimmer customization used on different backgrounds.
+   */
+  shimmerColors?: IShimmerColors;
 }
 
 /**
- * Shimmer Elements Interface
+ * Shimmer Elements Interface representing all common properties between Gap, Circle and Line.
  */
 export interface IShimmerElement {
   /**
+   * Represents the possible type of the shimmer elements: Gap, Circle, Line.
    * Required for every element you intend to use.
    */
   type: ShimmerElementType;
 
   /**
-   * The height of the element (ICircle, ILine, IGap) in pixels.
+   * Sets the height of the element (ICircle, ILine, IGap) in pixels.
    * Read more details for each specific element.
    */
   height?: number;
 
   /**
-   * The width value of the element (ILine, IGap) in pixels.
+   * Sets the width value of the element (ILine, IGap) in pixels.
    * Read more details for each specific element.
    */
   width?: number | string;
 
   /**
-   * The vertical alignemt of the element (ICircle, ILine).
+   * Sets vertical alignment of the element (ICircle, ILine).
    * @defaultvalue center
    */
   verticalAlign?: 'top' | 'center' | 'bottom';
 }
 
+/**
+ * Line element interface
+ */
 export interface ILine extends IShimmerElement {
   /**
    * Sets the height of the shimmer line in pixels.
@@ -99,6 +108,9 @@ export interface ILine extends IShimmerElement {
   width?: number | string;
 }
 
+/**
+ * Circle element interface
+ */
 export interface ICircle extends IShimmerElement {
   /**
    * Sets the height of the shimmer circle in pixels.
@@ -108,6 +120,9 @@ export interface ICircle extends IShimmerElement {
   height?: number;
 }
 
+/**
+ * Gap element interface
+ */
 export interface IGap extends IShimmerElement {
   /**
    * Sets the height of the shimmer gap in pixels.
@@ -122,20 +137,49 @@ export interface IGap extends IShimmerElement {
   width?: number | string;
 }
 
+/**
+ * Defines props needed to construct styles. This represents the simplified set of immutable things which control the class names.
+ */
 export interface IShimmerStyleProps {
+  /** Boolean flag to trigger fadeIn/fadeOut transition animation when content is loaded. */
   isDataLoaded?: boolean;
+
+  /** Optional CSS class name for the component attached to the root stylable area. */
   className?: string;
+
+  /** Theme provided by High-Order Component. */
   theme: ITheme;
+
+  /** Interval in milliseconds for the adeIn/fadeOut transition animation. */
   transitionAnimationInterval?: number;
+
+  /** Color to be used as the main background color of Shimmer when not animating. */
+  shimmerColor?: string;
+
+  /** Tip color of the shimmer wave which gradually gets from and to `shimmerColor`. */
+  shimmerWaveColor?: string;
 }
 
+/**
+ * Represents the stylable areas of the control.
+ */
 export interface IShimmerStyles {
+  /** Refers to the root wrapper element. */
   root?: IStyle;
+
+  /** Refers to wrapper element of the shimmer animation only. */
   shimmerWrapper?: IStyle;
+
+  /** Refers to wrapper element of the children only. */
   dataWrapper?: IStyle;
+
+  /** Styles for the hidden helper element to aid with screen readers. */
   screenReaderText?: IStyle;
 }
 
+/**
+ * Describes the possible types for shimmer elements used.
+ */
 export enum ShimmerElementType {
   /**
    * Line element type
@@ -153,6 +197,9 @@ export enum ShimmerElementType {
   gap = 3
 }
 
+/**
+ * Describes the default heights for shimmer elements when omitted in implementation.
+ */
 export enum ShimmerElementsDefaultHeights {
   /**
    * Default height of the line element when not provided by user: 16px
@@ -168,4 +215,27 @@ export enum ShimmerElementsDefaultHeights {
    * Default height of the circle element when not provided by user: 24px
    */
   circle = 24
+}
+
+/**
+ * Interface describing the possible color customizations of Shimmer.
+ */
+export interface IShimmerColors {
+  /**
+   * Defines the main background color which is the color you see when the wave is not animating.
+   * @defaultvalue theme.palette.neutralLight
+   */
+  shimmer?: string;
+
+  /**
+   * Defines the tip color of the shimmer wave which gradually gets from and to `shimmer` color.
+   * @defaultvalue theme.palette.neutralLighter
+   */
+  shimmerWave?: string;
+
+  /**
+   * Defines the background color of the space in between and around shimmer elements (borders, gaps and rounded corners).
+   * @defaultvalue theme.palette.white
+   */
+  background?: string;
 }