+++ packages/office-ui-fabric-react/src/components/Shimmer/examples/Shimmer.CustomElements.Example.tsx
@@ -1,7 +1,7 @@
 // @codepen
 
 import * as React from 'react';
-import { Shimmer, ShimmerElementsGroup, ShimmerElementType as ElemType } from 'office-ui-fabric-react/lib/Shimmer';
+import { Shimmer, ShimmerElementsGroup, ShimmerElementType } from 'office-ui-fabric-react/lib/Shimmer';
 import { Fabric } from 'office-ui-fabric-react/lib/Fabric';
 import { mergeStyles } from 'office-ui-fabric-react/lib/Styling';
 
@@ -37,14 +37,17 @@ export class ShimmerCustomElementsExample extends React.Component<{}, {}> {
         style={{ display: 'flex' }}
       >
         <ShimmerElementsGroup
-          shimmerElements={[{ type: ElemType.line, width: 40, height: 40 }, { type: ElemType.gap, width: 10, height: 40 }]}
+          shimmerElements={[
+            { type: ShimmerElementType.line, width: 40, height: 40 },
+            { type: ShimmerElementType.gap, width: 10, height: 40 }
+          ]}
         />
         <ShimmerElementsGroup
           flexWrap={true}
           shimmerElements={[
-            { type: ElemType.line, width: 300, height: 10 },
-            { type: ElemType.line, width: 200, height: 10 },
-            { type: ElemType.gap, width: 100, height: 20 }
+            { type: ShimmerElementType.line, width: 300, height: 10 },
+            { type: ShimmerElementType.line, width: 200, height: 10 },
+            { type: ShimmerElementType.gap, width: 100, height: 20 }
           ]}
         />
       </div>
@@ -57,13 +60,15 @@ export class ShimmerCustomElementsExample extends React.Component<{}, {}> {
         // tslint:disable-next-line:jsx-ban-props
         style={{ display: 'flex' }}
       >
-        <ShimmerElementsGroup shimmerElements={[{ type: ElemType.circle, height: 40 }, { type: ElemType.gap, width: 10, height: 40 }]} />
+        <ShimmerElementsGroup
+          shimmerElements={[{ type: ShimmerElementType.circle, height: 40 }, { type: ShimmerElementType.gap, width: 10, height: 40 }]}
+        />
         <ShimmerElementsGroup
           flexWrap={true}
           shimmerElements={[
-            { type: ElemType.line, width: 400, height: 10 },
-            { type: ElemType.gap, width: 100, height: 20 },
-            { type: ElemType.line, width: 500, height: 10 }
+            { type: ShimmerElementType.line, width: 400, height: 10 },
+            { type: ShimmerElementType.gap, width: 100, height: 20 },
+            { type: ShimmerElementType.line, width: 500, height: 10 }
           ]}
         />
       </div>
@@ -78,31 +83,36 @@ export class ShimmerCustomElementsExample extends React.Component<{}, {}> {
       >
         <ShimmerElementsGroup
           width={'90px'}
-          shimmerElements={[{ type: ElemType.line, height: 80, width: 80 }, { type: ElemType.gap, width: 10, height: 80 }]}
+          shimmerElements={[
+            { type: ShimmerElementType.line, height: 80, width: 80 },
+            { type: ShimmerElementType.gap, width: 10, height: 80 }
+          ]}
         />
         <div
           // tslint:disable-next-line:jsx-ban-props
           style={{ display: 'flex', flexWrap: 'wrap', width: '100%' }}
         >
-          <ShimmerElementsGroup shimmerElements={[{ type: ElemType.circle, height: 40 }, { type: ElemType.gap, width: 10, height: 40 }]} />
+          <ShimmerElementsGroup
+            shimmerElements={[{ type: ShimmerElementType.circle, height: 40 }, { type: ShimmerElementType.gap, width: 10, height: 40 }]}
+          />
           <ShimmerElementsGroup
             flexWrap={true}
             width={'calc(100% - 50px)'}
             shimmerElements={[
-              { type: ElemType.line, width: '90%', height: 10 },
-              { type: ElemType.gap, width: '10%', height: 20 },
-              { type: ElemType.line, width: '100%', height: 10 }
+              { type: ShimmerElementType.line, width: '90%', height: 10 },
+              { type: ShimmerElementType.gap, width: '10%', height: 20 },
+              { type: ShimmerElementType.line, width: '100%', height: 10 }
             ]}
           />
           <ShimmerElementsGroup
             flexWrap={true}
             width={'100%'}
             shimmerElements={[
-              { type: ElemType.line, width: '80%', height: 10, verticalAlign: 'bottom' },
-              { type: ElemType.gap, width: '20%', height: 20 },
-              { type: ElemType.line, width: '40%', height: 10, verticalAlign: 'bottom' },
-              { type: ElemType.gap, width: '2%', height: 20 },
-              { type: ElemType.line, width: '58%', height: 10, verticalAlign: 'bottom' }
+              { type: ShimmerElementType.line, width: '80%', height: 10, verticalAlign: 'bottom' },
+              { type: ShimmerElementType.gap, width: '20%', height: 20 },
+              { type: ShimmerElementType.line, width: '40%', height: 10, verticalAlign: 'bottom' },
+              { type: ShimmerElementType.gap, width: '2%', height: 20 },
+              { type: ShimmerElementType.line, width: '58%', height: 10, verticalAlign: 'bottom' }
             ]}
           />
         </div>