+++ packages/office-ui-fabric-react/src/components/SwatchColorPicker/ColorPickerGridCell.tsx
@@ -3,7 +3,7 @@ import { ColorPickerGridCellBase } from './ColorPickerGridCell.base';
 import { IColorPickerGridCellProps, IColorPickerGridCellStyleProps, IColorPickerGridCellStyles } from './ColorPickerGridCell.types';
 import { getStyles } from './ColorPickerGridCell.styles';
 
-export const ColorPickerGridCell: (props: IColorPickerGridCellProps) => JSX.Element = styled<
+export const ColorPickerGridCell: React.StatelessComponent<IColorPickerGridCellProps> = styled<
   IColorPickerGridCellProps,
   IColorPickerGridCellStyleProps,
   IColorPickerGridCellStyles